Make more private
